+++ data/src/main/java/com/wakeup/data/di/DataSourceModule.kt
@@ -4,6 +4,8 @@ import com.wakeup.data.source.local.globe.GlobeLocalDataSource
 import com.wakeup.data.source.local.globe.GlobeLocalDataSourceImpl
 import com.wakeup.data.source.local.moment.MomentLocalDataSource
 import com.wakeup.data.source.local.moment.MomentLocalDataSourceImpl
+import com.wakeup.data.source.local.ref.RefLocalDataSource
+import com.wakeup.data.source.local.ref.RefLocalDataSourceImpl
 import com.wakeup.data.source.remote.placeSearch.PlaceSearchRemoteDataSource
 import com.wakeup.data.source.remote.placeSearch.PlaceSearchRemoteDataSourceImpl
 import dagger.Binds
@@ -29,4 +31,9 @@ interface DataSourceModule {
     fun bindGlobeLocalDataSource(
         globeLocalDataSourceImpl: GlobeLocalDataSourceImpl,
     ): GlobeLocalDataSource
+
+    @Binds
+    fun bindRefLocalDataSource(
+        refLocalDataSourceImpl: RefLocalDataSourceImpl,
+    ): RefLocalDataSource
 }
