+++ data/src/main/java/com/wakeup/data/source/local/globe/GlobeLocalDataSourceImpl.kt
@@ -4,7 +4,6 @@ import com.wakeup.data.database.dao.GlobeDao
 import com.wakeup.data.database.entity.GlobeEntity
 import com.wakeup.data.database.entity.MomentWithGlobesAndPictures
 import kotlinx.coroutines.flow.Flow
-import timber.log.Timber
 import javax.inject.Inject
 
 class GlobeLocalDataSourceImpl @Inject constructor(
@@ -20,8 +19,11 @@ class GlobeLocalDataSourceImpl @Inject constructor(
     }
 
     override suspend fun updateGlobe(globe: GlobeEntity) {
-        Timber.d("updateGlobe: $globe")
         globeDao.updateGlobe(globe)
+    }
+
+    override suspend fun getGlobeId(globeName: String): Long {
+        return globeDao.getGlobeIdByName(globeName)
     }
 
     override fun getGlobes(): Flow<List<GlobeEntity>> {
